Fixed bug where notif quick answer entryId was -1

+++ app/src/main/java/com/lubenard/oring_reminder/Utils.java
@@ -99,7 +99,7 @@ public class Utils {
      * @param content notification body
      * @param drawable drawable icon
      */
-    public static void sendNotificationWithQuickAnswer(Context context, String title, String content, int drawable, int entryId) {
+    public static void sendNotificationWithQuickAnswer(Context context, String title, String content, int drawable, long entryId) {
         // First let's create the intent
         PendingIntent pi = PendingIntent.getActivity(context, 1, new Intent(context, MainActivity.class), PendingIntent.FLAG_UPDATE_CURRENT);
